folder page, shared page 분할
