+++ weeklymission2/src/styles/folderHeader.css
@@ -1,0 +1,43 @@
+.linkInputArea{
+  margin: 0 10rem;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.addLinks{
+  padding: 2rem 4.5rem;
+  border-color: #6D6AFF;
+  border-radius: 1rem;
+  margin: 4rem 0 6rem;
+  width: 100%;
+
+}
+.linkIcon{
+  position: absolute;
+  left: 2rem;
+  bottom: 7.9rem;
+}
+.addBtn{
+  background-image: linear-gradient(91deg, var(--primary), #6AE3FE 101.84%);
+  border-radius: 0.4rem;
+  color: #f5f5f5;
+  font-weight: 600;
+  position: absolute;
+  top: 5.4rem;
+  right: 2rem;
+  font-size: 1rem;
+  padding: 1rem;
+}
+
+@media (max-width: 1199px) {
+  .linkInputArea{
+    margin: 0 5rem;
+  }
+}
+
+@media (max-width: 767px) {
+  .linkInputArea{
+    margin: 0 3rem;
+  }
+}
