+++ weeklymission2/src/styles/sharedHeader.css
@@ -1,0 +1,25 @@
+header{
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.folderProfile{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+.folderImg{
+  margin-top: 2rem;
+  height: 7rem;
+  width: 7rem;
+}
+.folderProfileName{
+  margin: 1rem;
+  font-size: 1.8rem;
+}
+.folderName{
+  font-size: 4rem;
+  font-weight: 600;
+  margin: 2rem 0 7rem;
+}
